Rename Highscores table and move data files to data directory
