+++ backend/backend/AimTrainerGame/Data/DotInfo.cs
@@ -1,7 +1,7 @@
 using System.Text.Json.Serialization;
 using backend.Utils;
 
-namespace backend.AimTrainerGame.Models;
+namespace backend.AimTrainerGame.Data;
 
 public struct DotInfo
 {
