+++ backend/backend/AimTrainerGame/Data/GameEndRequest.cs
@@ -1,4 +1,4 @@
-namespace backend.AimTrainerGame.Models;
+namespace backend.AimTrainerGame.Data;
 
 public record GameEndRequest(string Username, int Score)
 {
