+++ backend/backend/AimTrainerGame/Data/GameStartRequest.cs
@@ -1,6 +1,6 @@
 using backend.AimTrainerGame.Settings;
 using backend.Utils;
 
-namespace backend.AimTrainerGame.Models;
+namespace backend.AimTrainerGame.Data;
 
 public record GameStartRequest(Difficulty difficulty, Vec2<int> screenSize);
