+++ backend/backend/AimTrainerGame/Data/GameStartResponse.cs
@@ -1,3 +1,3 @@
-namespace backend.AimTrainerGame.Models;
+namespace backend.AimTrainerGame.Data;
 
 public record GameStartResponse(List<DotInfo> dotInfos, int dotCount, int timeToLive);
