+++ backend/backend/AimTrainerGame/Models/AimTrainerGameHighscoreContext.cs
@@ -4,5 +4,5 @@ namespace backend.AimTrainerGame.Models;
 
 public class AimTrainerGameHighscoreContext(DbContextOptions<AimTrainerGameHighscoreContext> options) : DbContext(options)
 {
-    public DbSet<Highscore> Highscores { get; set; }
+    public DbSet<Highscore> AimTrainerGameHighscores { get; set; }
 }
